requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
supabase==2.3.0
python-dotenv==1.0.0
paramiko==3.4.0
xhtml2pdf==0.2.13
Werkzeug==3.0.1
email-validator==2.1.0
